+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.5] - 2026-07-06
+
+### Fixed
+
+- Fixed all criteria displaying an additional empty `player:null` condition on 1.20.
+- Fixed criteria triggers not autofilling until you start typing.
+
 ## [4.0.4] - 2026-06-26
 
 ### Fixed